feat&refactor: 增加console显隐选项、增加无console打包、修复x64类型安全问题
1. 增加无Console打包
2. 增加默认Console显示选项，并可以用命令行参数重载
3. 重构异常处理机制
4. 重构一些数据类型，确保x64运行安全
5. 分离metadata的读取代码，优化代码格式

+++ Project1/archive.cpp
@@ -126,7 +126,7 @@ bool inflate_file(std::string fileIn, std::string fileOut)
 }
 
 // 解压到文件
-bool inflate_to_file(std::fstream& fileIn, uint64_t offset, uint64_t length, std::string fileOut)
+bool inflate_to_file(std::fstream& fileIn, size_t offset, size_t length, std::string fileOut)
 {
     std::fstream fout(fileOut, std::fstream::out | std::fstream::binary | std::fstream::trunc);
